MAINT: Get rid of illegal template virtual functions
C++ doesn't allow virtual template functions since, when they're
declared in the headers, it won't know which specific implementations of
them need to be created.  This is a problem because most C++ compilers
use vtables on each class to keep track of virtual function resolution,
and creating the vtable means you need to know exactly how many entries
to create.